python_service/requirements.txt
flask
pillow
pydub
requests
python-dotenv
numpy 